Had to make template reader not throw exceptions
This sucks because now the GUI won't have a way of knowing if
TemplateReader was successful or not, but we couldn't get code coverage
with it. So I removed those exceptions and just had it return the empty
string like it used to.

Also, now that there's stuff in the lib directory, the build script
needed to know to consider that when running findbugs.