Fixed line ending issue
Cody was a dumbass and didn't use the built-in java stuff for reading and
writing strings invloving new line characters. This change made everything not
system independent anymore which cause the tests to not pass on some of our
machines. This place where I broke the code has been patched by translating
all the of the newline characters to unix newline characters.